+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Fix PDBj structure data URL
+- Improve logic when to cull in renderer
 
 ## [v4.11.0] - 2025-01-26