+++ CHANGELOG.md
@@ -11,8 +11,8 @@ Note that since we don't clearly distinguish between a public and private interf
 - Better IH/M support in MolViewSpec:
   - Support `coarse` components
   - Support `spacefill` representation
-  - Support `element_granularity` in component expressions
-  - If no color theme is specified explicitly, use Mol*'s defaults instead
+  - Support for `custom.molstar_use_default_coloring` property on Color node.
+  - Use `atom.ihm.has-seq-id` for matching `label_seq_id` locations to support querying coarse elements.
 
 ## [v4.11.0] - 2025-01-26