+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Fix PDBj structure data URL
+- Improve logic when to cull in renderer
+
+## [v4.11.0] - 2025-01-26
+
 - Fix for tubular helices issue (Fixes #1422)
 - Volume UI improvements
     - Render all volume entries instead of selecting them one-by-one